Fix fetching gems with platforms (#33)
`#full_name` includes the platform, whereas the ad-hoc
`#name`/`#version` combo does not.

Note that if a gem (such as facter) has platform and non-platform
versions, bundix will still always put the non-platform version in
gemset.nix. I don't think this is fixable without some work on the
format of gemset.nix (which would mean changes in nixpkgs too),
because the correct solution would be for _both_ to be included.

Fixes #14.